Fix mypy issues

+++ baybe/targets/numerical.py
@@ -231,11 +231,11 @@ class NumericalTarget(Target, SerialMixin):
         Returns:
             The target with applied quadratic matching transformation.
         """
-        return NumericalTarget.match_power(name, match_value, exponent=2.0)
+        return NumericalTarget.match_power(name, match_value, exponent=2)
 
     @classmethod
     def match_power(
-        cls, name: str, match_value: float, exponent: float
+        cls, name: str, match_value: float, exponent: int
     ) -> NumericalTarget:
         """Create a target to match a given value using a power transformation.
